pyproject.toml
[project]
name = "ia-tools"
version = "0.1.0"
description = "Centralized AI toolbox for development teams"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
ia-memory = "scripts.memory.query:main"
ia-eval = "evals.runners.run_eval:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.5.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
